+++ dwengo_backend/routes/progress/studentProgressRoutes.ts
@@ -5,12 +5,9 @@ import {
   updateProgress,
   getTeamProgressStudent,
   getStudentAssignmentProgress,
-  getTeamProgressTeacher,
-  getAssignmentAverageProgress,
   getStudentLearningPathProgress,
-} from "../controllers/progressController";
-import { protectStudent } from "../middleware/studentAuthMiddleware";
-import { protectTeacher } from "../middleware/teacherAuthMiddleware";
+} from "../../controllers/progressController";
+import { protectStudent } from "../../middleware/studentAuthMiddleware";
 
 const router: Router = express.Router();
 
@@ -28,7 +25,7 @@ const router: Router = express.Router();
  * @access  Student
  */
 router.post(
-  "/student/learningObject/:learningObjectId",
+  "/learningObject/:learningObjectId",
   protectStudent,
   createProgress
 );
@@ -41,7 +38,7 @@ router.post(
  * @access  Student
  */
 router.get(
-  "/student/learningObject/:learningObjectId",
+  "/learningObject/:learningObjectId",
   protectStudent,
   getStudentProgress
 );
@@ -54,7 +51,7 @@ router.get(
  * @access  Student
  */
 router.patch(
-  "/student/learningObject/:learningObjectId",
+  "/learningObject/:learningObjectId",
   protectStudent,
   updateProgress
 );
@@ -69,7 +66,7 @@ router.patch(
  * @param  teamid: number
  * @access  Student
  */
-router.get("/student/team/:teamid", protectStudent, getTeamProgressStudent);
+router.get("/team/:teamid", protectStudent, getTeamProgressStudent);
 
 /**
  * @route   GET /progress/student/assignment/:assignmentId
@@ -80,7 +77,7 @@ router.get("/student/team/:teamid", protectStudent, getTeamProgressStudent);
  * @access  Student
  */
 router.get(
-  "/student/assignment/:assignmentId",
+  "/assignment/:assignmentId",
   protectStudent,
   getStudentAssignmentProgress
 );
@@ -94,40 +91,9 @@ router.get(
  * @access  Student
  */
 router.get(
-  "/student/learningPath/:learningPathId/",
+  "/learningPath/:learningPathId/",
   protectStudent,
   getStudentLearningPathProgress
-);
-/**
- * ===========================
- *          TEACHER
- * ===========================
- */
-
-/**
- * @route   GET /progress/teacher/team/:teamid
- * @desc    Haal de voortgang van een team op voor een specifieke opdracht.
- *          Dit betekent dat je de progressie van alle teamleden bekijkt en bepaalt
- *          wat de verste vooruitgang is die iemand binnen het team heeft bereikt.
- *          Geeft een percentage terug van het verste dat iemand binnen een team is geraakt.
- *          Je rekent het percentage uit van het verste leerobject dat is bereikt binnen het leerpad van alle leerobjecten binnen dat leerpad
- * @param  teamid: number
- * @access  Teacher
- */
-router.get("/teacher/team/:teamid", protectTeacher, getTeamProgressTeacher);
-
-/**
- * @route   GET /progress/teacher/assignment/:assignmentId/average
- * @desc    Bereken de gemiddelde voortgang van een klas bij een opdracht.
- *          De verste vooruitgang van elk team of student wordt verzameld en
- *          er wordt een gemiddelde berekend.
- * @param  assignmentId: number
- * @access  Teacher
- */
-router.get(
-  "/teacher/assignment/:assignmentId/average",
-  protectTeacher,
-  getAssignmentAverageProgress
 );
 
 export default router;
